+++ AES/AES.h
@@ -15,9 +15,9 @@ private:
 	static const unsigned char SBOX[16][16]; //represent the SBOX table for AES
 	static const unsigned char INVSBOX[16][16]; //represent the inverse SBOX table for AES
 	static const unsigned char GaloisMult[15][256]; //represents the Galois multiplication tables for AES
-	static const int Nb = 4; //number of columns in the state (always 4 for AES)
-	static int Nk; //number of 32-bit words in the key 
-	static int Nr; //number of rounds (AES-128 has 10 rounds, AES-192 has 12 rounds, AES-256 has 14 rounds)
+	static const size_t Nb = 4; //number of columns in the state (always 4 for AES)
+	static size_t Nk; //number of 32-bit words in the key 
+	static size_t Nr; //number of rounds (AES-128 has 10 rounds, AES-192 has 12 rounds, AES-256 has 14 rounds)
 	static const size_t BlockSize = Nb * Nb; //represents the size of AES block that is always 16 bytes (128-bit)
 
 protected:
